+++ src/main/java/com/kyouseipro/neo/sales/order/repository/WorkContentRepository.java
@@ -20,7 +20,7 @@ public class WorkContentRepository {
      * 0件の場合は空リストを返す。
      * @return 取得したリストを返す
      */
-    public List<WorkContentEntity> findAllByOrderId(int id, String editor) {
+    public List<WorkContentEntity> findAllByOrderId(Long id, String editor) {
         String sql = WorkContentSqlBuilder.buildFindAllByOrderId();
 
         return sqlRepository.queryList(
